+++ RiceRevolution/server.py
@@ -8,6 +8,11 @@ app = Flask(__name__)
 def home():
    return render_template('home.html')
 
+# Varieties page
+@app.route('/varieties')
+def varieties():
+   return render_template('varieties.html')
+
 
 if __name__ == '__main__':
    app.run(port=8000, debug = True)
