+++ RiceRevolution/static/varieties.js
@@ -1,0 +1,10 @@
+import * as functions from './functions.js';
+
+$(document).ready(function() {
+    renderVarietiesHome();
+});
+
+function renderVarietiesHome() {
+    // Render global margins
+    functions.renderMargins();
+}
