+++ RiceRevolution/templates/layout.html
@@ -31,7 +31,7 @@
       <!-- TODO: Add links -->
       <a class="navbar-brand" href="/">Rice Revolution</a>
       <a class="navbar-brand" href="/">Home</a>
-      <a class="navbar-brand" href="/">Varieties</a>
+      <a class="navbar-brand" href="/varieties">Varieties</a>
       <a class="navbar-brand" href="/">No Rice Cooker</a>
       <a class="navbar-brand" href="/">Rice Cooker</a>
       <a class="navbar-brand" href="/">Quiz</a>
