+++ RiceRevolution/templates/varieties.html
@@ -1,0 +1,13 @@
+{% extends "layout.html" %}
+
+{% block content %}
+
+<script type="module" src="{{ url_for('static', filename = 'varieties.js') }}"></script>
+
+<script>
+    console.log('varities')
+</script>
+
+<div class="container" id="main"></div>
+
+{% endblock %}
